requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
setuptools==67.8.0
alive-progress==3.1.2
python-dotenv==1.0.0
python-ffmpeg==2.0.4
ffmpeg==1.4
ffmpeg-python==0.2.0
keyboard==0.13.5
torch==2.0.1
Pillow==9.5.0
drive==0.4.1
requests~=2.31.0